refactor(server): refact cli with cac;use esm format instead
