+++ .changeset/gentle-shirts-marry.md
@@ -5,3 +5,4 @@
 replace compiler with generator to provide better scalability
 loader can read config from config file, it's more flexible to control loader function
 it's a big change for both runtime and compile sysmtem
+refact cli with `cac`; use esm format instead
