+++ packages/server/package.json
@@ -6,7 +6,7 @@
   "module": "dist/index.mjs",
   "types": "dist/index.d.ts",
   "bin": {
-    "phecda": "bin/cli.js"
+    "phecda": "bin/cli.mjs"
   },
   "scripts": {
     "dev": "tsup --watch",
